refactor: :recycle: refatorando as pastas e aplicando camelCase no nome dos arquivos
