--- src/Bank/conta/conta.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ContaService } from './conta.service';
-import { ContaController } from './conta.controller';
-
-@Module({
-  providers: [ContaService],
-  controllers: [ContaController]
-})
-export class ContaModule {}
