--- src/Bank/gerente/gerente.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { GerenteService } from './gerente.service';
-import { GerenteController } from './gerente.controller';
-import { ClienteModule } from '../cliente/cliente.module';
-
-@Module({
-  imports: [ClienteModule], 
-  providers: [GerenteService],
-  controllers: [GerenteController],
-})
-export class GerenteModule {}
