+++ src/modules/contaModule.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { ContaService } from '../services/contaService';
+import { ContaController } from '../controllers/contaController';
+
+@Module({
+  providers: [ContaService],
+  controllers: [ContaController]
+})
+export class ContaModule { }
